Updated map, added Skills and Interface classes

+++ src/com/pathfinder/EntityPlayer.java
@@ -15,7 +15,7 @@ public class EntityPlayer extends EntityMob
 		getInput();
 		move();
 		Game.camera.center(this);
-		System.out.println("Location: " + (int) x/128 + "," + (int) y/128);
+		System.out.println("Location: " + (double) x/128 + "," + (double) y/128);
 		System.out.println("Movement Booleans: " + Display.up + "," + Display.down + "," + Display.right + "," + Display.left);
 	}
